flask/requirements.txt
Flask==2.2.5
Werkzeug==2.2.2
MarkupSafe==2.1.1
twilio==9.6.0
python-dotenv==1.1.0
